requirements.txt
youtube-transcript-api
langchain
langchain_text_splitters
sentence-transformers
torch
faiss-cpu
ollama
